+++ utils/memusage.h
@@ -1,13 +1,15 @@
 #ifndef MEMUSAGE_H_
 #define MEMUSAGE_H_
 
+#include <stddef.h>
+
 #ifdef	__cplusplus
 extern "C" {
 #endif
 
-extern long Memusage (void);
-extern long Memusage2 (void);
-extern long PeakMemusage (void);
+extern size_t Memusage (void);
+extern size_t Memusage2 (void);
+extern size_t PeakMemusage (void);
 
 #ifdef	__cplusplus
 }
